+++ test/SILOptimizer/definite-init-wrongscope.swift
@@ -1,5 +1,5 @@
 // RUN: %target-swift-frontend -primary-file %s -Onone -emit-sil -Xllvm \
-// RUN:   -sil-print-after=definite-init -Xllvm \
+// RUN:   -sil-print-after=raw-sil-inst-lowering -Xllvm \
 // RUN:   -sil-print-only-functions=$S3del1MC4fromAcA12WithDelegate_p_tKcfc \
 // RUN:   -Xllvm -sil-print-debuginfo -o /dev/null 2>&1 | %FileCheck %s
