+++ test/SILOptimizer/di-conditional-destroy-scope.swift
@@ -1,5 +1,5 @@
 // RUN: %target-swift-frontend -emit-sil %s -Onone -Xllvm \
-// RUN:   -sil-print-after=definite-init -Xllvm \
+// RUN:   -sil-print-after=raw-sil-inst-lowering -Xllvm \
 // RUN:   -sil-print-only-functions=$S2fs36RecursibleDirectoryContentsGeneratorC4path10fileSystemAcA12AbsolutePathV_AA04FileH0_ptKc33_F8B132991B28208F48606E87DC165393Llfc \
 // RUN:   -Xllvm -sil-print-debuginfo -o /dev/null 2>&1 | %FileCheck %s
